cloudfunction/process_audio/requirements.txt
functions-framework==3.*
flask
vertexai
urllib3
google-cloud-storage
google-cloud-speech
google-cloud-texttospeech
pydub